+++ CHANGELOG.md
@@ -12,12 +12,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Renamed feature `matrix_graph` to `matrix-graph`
 - Renamed feature `stable_graph` to `stable-graph`
 - `CSR` is now gated behind the `csr` feature
-- `AdjacencyMatrix` is now gated behind the `adjacency-matrix` feature
+- `AdjacencyList` is now gated behind the `adjacency-matrix` feature
 - `dot` format is now gated behind the `dot` feature
 - Replaced rendering of `dot` graphs with `dot` crate, instead of internal
   implementation. Previous configuration options are no longer available.
 
 [//]: # (TODO: comtemplate moving to GraphViz-Rust crate)
+[//]: # (TODO: should `AdjacencyList` and `AdjacencyMatrix` be in one crate, how to call that crate? `petgraph-adjacency`, I mean you could think of it as a sparse matrix, so it's fine?)
 
 ### Changed